movies_admin/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-debug-toolbar==4.3.0
django-split-settings==1.3.1
flake8==6.1.0
mccabe==0.7.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
uWSGI==2.0.25.1
